Move base filter and order into their namespaces
