+++ app/models/queries/group_bys/base.rb
