+++ app/models/queries/individual_principals/orders/group_order.rb
@@ -27,7 +27,7 @@
 # See COPYRIGHT and LICENSE files for more details.
 #++
 
-class Queries::IndividualPrincipals::Orders::GroupOrder < Queries::BaseOrder
+class Queries::IndividualPrincipals::Orders::GroupOrder < Queries::Orders::Base
   self.model = Principal
 
   def self.key
