+++ app/models/queries/members/orders/default_order.rb
@@ -28,7 +28,7 @@
 # See COPYRIGHT and LICENSE files for more details.
 #++
 
-class Queries::Members::Orders::DefaultOrder < Queries::BaseOrder
+class Queries::Members::Orders::DefaultOrder < Queries::Orders::Base
   self.model = Member
 
   def self.key
