+++ app/models/queries/notifications/orders/reason_order.rb
@@ -28,7 +28,7 @@
 # See COPYRIGHT and LICENSE files for more details.
 #++
 
-class Queries::Notifications::Orders::ReasonOrder < Queries::BaseOrder
+class Queries::Notifications::Orders::ReasonOrder < Queries::Orders::Base
   self.model = Notification
 
   def self.key
