+++ app/models/queries/projects/orders/latest_activity_at_order.rb
@@ -28,7 +28,7 @@
 # See COPYRIGHT and LICENSE files for more details.
 #++
 
-class Queries::Projects::Orders::LatestActivityAtOrder < Queries::BaseOrder
+class Queries::Projects::Orders::LatestActivityAtOrder < Queries::Orders::Base
   self.model = Project
 
   def self.key
