+++ app/models/queries/projects/orders/name_order.rb
@@ -28,7 +28,7 @@
 # See COPYRIGHT and LICENSE files for more details.
 #++
 
-class Queries::Projects::Orders::NameOrder < Queries::BaseOrder
+class Queries::Projects::Orders::NameOrder < Queries::Orders::Base
   self.model = Project
 
   def self.key
