+++ app/models/queries/projects/orders/project_status_order.rb
@@ -28,7 +28,7 @@
 # See COPYRIGHT and LICENSE files for more details.
 #++
 
-class Queries::Projects::Orders::ProjectStatusOrder < Queries::BaseOrder
+class Queries::Projects::Orders::ProjectStatusOrder < Queries::Orders::Base
   self.model = Project
 
   def self.key
